app/requirements.txt
fastapi
uvicorn
openai
langchain_openai
python-dotenv
sqlalchemy
pydantic_settings
pymysql
aiomysql